Further consolidate into main.ts
